Add possibility to set default currency per organisation

+++ backend/internal/service/db_service/queries/delete_vat.sql
@@ -1,4 +1,4 @@
 DELETE FROM vats
 WHERE
     id = ?
-    AND organisation_id = get_current_organisation(?)
+    AND organisation_id = get_current_user_organisation_id(?)
